+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -7003,10 +7003,6 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             unsupported_conditions=[
                 ("dummy_na", True),
                 ("drop_first", True),
-                (
-                    lambda args: args.get("dtype") is not None,
-                    "get_dummies with non-default dtype parameter is not supported yet in Snowpark pandas.",
-                ),
             ]
         ),
     )
@@ -7049,9 +7045,9 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         """
         self._raise_not_implemented_error_for_timedelta()
 
-        if dummy_na is True or drop_first is True or dtype is not None:
+        if dummy_na is True or drop_first is True:
             ErrorMessage.not_implemented(
-                "get_dummies with non-default dummy_na, drop_first, and dtype parameters"
+                "get_dummies with non-default dummy_na or drop_first parameters"
                 + " is not supported yet in Snowpark pandas."
             )
         if columns is None:
@@ -7095,6 +7091,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             columns=columns,
             prefixes=prefix,
             prefix_sep=prefix_sep,
+            dtype=dtype,
         )
         query_compiler = SnowflakeQueryCompiler(result_internal_frame)
 
@@ -7807,6 +7804,34 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
     @snowpark_pandas_type_immutable_check
     def rename(
+        self,
+        *,
+        index_renamer: Optional[Renamer] = None,
+        columns_renamer: Optional[Renamer] = None,
+        # TODO: SNOW-800889 handle level is hashable
+        level: Optional[Union[Hashable, int]] = None,
+        errors: Optional[IgnoreRaise] = "ignore",
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _rename_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = self._relaxed_query_compiler._rename_internal(
+                index_renamer=index_renamer,
+                columns_renamer=columns_renamer,
+                level=level,
+                errors=errors,
+            )
+        qc = self._rename_internal(
+            index_renamer=index_renamer,
+            columns_renamer=columns_renamer,
+            level=level,
+            errors=errors,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _rename_internal(
         self,
         *,
         index_renamer: Optional[Renamer] = None,
@@ -11584,6 +11609,77 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         frame_is_df_and_item_is_series: bool = False,
     ) -> "SnowflakeQueryCompiler":
         """
+        Wrapper around _set_2d_labels_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if (
+            self._relaxed_query_compiler is not None
+            and (
+                not isinstance(index, SnowflakeQueryCompiler)
+                or index._relaxed_query_compiler is not None
+            )
+            and (
+                not isinstance(columns, SnowflakeQueryCompiler)
+                or columns._relaxed_query_compiler is not None
+            )
+            and (
+                not isinstance(item, SnowflakeQueryCompiler)
+                or item._relaxed_query_compiler is not None
+            )
+        ):
+            new_index = index
+            if isinstance(index, SnowflakeQueryCompiler):
+                new_index = index._relaxed_query_compiler
+            new_columns = columns
+            if isinstance(columns, SnowflakeQueryCompiler):
+                new_columns = columns._relaxed_query_compiler
+            new_item = item
+            if isinstance(item, SnowflakeQueryCompiler):
+                new_item = item._relaxed_query_compiler
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._set_2d_labels_internal(
+                    index=new_index,
+                    columns=new_columns,
+                    item=new_item,
+                    matching_item_columns_by_label=matching_item_columns_by_label,
+                    matching_item_rows_by_label=matching_item_rows_by_label,
+                    index_is_bool_indexer=index_is_bool_indexer,
+                    deduplicate_columns=deduplicate_columns,
+                    frame_is_df_and_item_is_series=frame_is_df_and_item_is_series,
+                )
+            )
+
+        qc = self._set_2d_labels_internal(
+            index=index,
+            columns=columns,
+            item=item,
+            matching_item_columns_by_label=matching_item_columns_by_label,
+            matching_item_rows_by_label=matching_item_rows_by_label,
+            index_is_bool_indexer=index_is_bool_indexer,
+            deduplicate_columns=deduplicate_columns,
+            frame_is_df_and_item_is_series=frame_is_df_and_item_is_series,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _set_2d_labels_internal(
+        self,
+        index: Union[Scalar, slice, "SnowflakeQueryCompiler"],
+        columns: Union[
+            "SnowflakeQueryCompiler",
+            tuple,
+            slice,
+            list,
+            "pd.Index",
+            np.ndarray,
+        ],
+        item: Union[Scalar, AnyArrayLike, "SnowflakeQueryCompiler"],
+        matching_item_columns_by_label: bool,
+        matching_item_rows_by_label: bool,
+        index_is_bool_indexer: bool,
+        deduplicate_columns: bool = False,
+        frame_is_df_and_item_is_series: bool = False,
+    ) -> "SnowflakeQueryCompiler":
+        """
         Create a new SnowflakeQueryCompiler with indexed columns and rows replaced by item.
 
         Args:
@@ -12983,6 +13079,32 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def drop(
+        self,
+        index: Optional[Sequence[Hashable]] = None,
+        columns: Optional[Sequence[Hashable]] = None,
+        level: Optional[Level] = None,
+        errors: Literal["raise", "ignore"] = "raise",
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _drop_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None and index is None:
+            relaxed_query_compiler = self._relaxed_query_compiler._drop_internal(
+                index=index,
+                columns=columns,
+                level=level,
+                errors=errors,
+            )
+        qc = self._drop_internal(
+            index=index,
+            columns=columns,
+            level=level,
+            errors=errors,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _drop_internal(
         self,
         index: Optional[Sequence[Hashable]] = None,
         columns: Optional[Sequence[Hashable]] = None,