+++ tests/integ/modin/test_faster_pandas.py
@@ -194,6 +194,35 @@ def test_read_filter_join_flag_disabled(session):
         assert_frame_equal(snow_result, native_result)
 
 
+@sql_count_checker(query_count=3)
+def test_drop(session):
+    # create tables
+    table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.create_dataframe(
+        native_pd.DataFrame([[2, True], [1, False], [3, False]], columns=["A", "B"])
+    ).write.save_as_table(table_name, table_type="temp")
+
+    # create snow dataframes
+    df = pd.read_snowflake(table_name)
+    snow_result = df.drop(columns=["B"])
+
+    # verify that the input dataframe has a populated relaxed query compiler
+    assert df._query_compiler._relaxed_query_compiler is not None
+    assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    # verify that the output dataframe also has a populated relaxed query compiler
+    assert snow_result._query_compiler._relaxed_query_compiler is not None
+    assert (
+        snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    )
+
+    # create pandas dataframes
+    native_df = df.to_pandas()
+    native_result = native_df.drop(columns=["B"])
+
+    # compare results
+    assert_frame_equal(snow_result, native_result)
+
+
 @pytest.mark.parametrize("func", ["isna", "isnull", "notna", "notnull"])
 @sql_count_checker(query_count=3)
 def test_isna_notna(session, func):
@@ -280,6 +309,35 @@ def test_isin_series(session):
 
     # compare results
     assert_frame_equal(snow_result, native_result, check_dtype=False)
+
+
+@sql_count_checker(query_count=3)
+def test_rename(session):
+    # create tables
+    table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.create_dataframe(
+        native_pd.DataFrame([[2, 12], [1, 11], [3, 13]], columns=["A", "B"])
+    ).write.save_as_table(table_name, table_type="temp")
+
+    # create snow dataframes
+    df = pd.read_snowflake(table_name)
+    snow_result = df.rename(columns={"A": "a", "B": "b"})
+
+    # verify that the input dataframe has a populated relaxed query compiler
+    assert df._query_compiler._relaxed_query_compiler is not None
+    assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    # verify that the output dataframe also has a populated relaxed query compiler
+    assert snow_result._query_compiler._relaxed_query_compiler is not None
+    assert (
+        snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    )
+
+    # create pandas dataframes
+    native_df = df.to_pandas()
+    native_result = native_df.rename(columns={"A": "a", "B": "b"})
+
+    # compare results
+    assert_frame_equal(snow_result, native_result)
 
 
 @sql_count_checker(query_count=3)
@@ -460,6 +518,87 @@ def test_sort_values(session):
     # create pandas dataframes
     native_df = df.to_pandas()
     native_result = native_df.sort_values(by="A")
+
+    # compare results
+    assert_frame_equal(snow_result, native_result)
+
+
+@sql_count_checker(query_count=3)
+def test_set_2d_labels_from_same_df(session):
+    # create tables
+    table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.create_dataframe(
+        native_pd.DataFrame([[2, 12], [1, 11], [3, 13]], columns=["A", "B"])
+    ).write.save_as_table(table_name, table_type="temp")
+
+    # create snow dataframes
+    df = pd.read_snowflake(table_name)
+    df["C"] = df["B"] + 1
+    snow_result = df
+
+    # verify that the input dataframe has a populated relaxed query compiler
+    assert df._query_compiler._relaxed_query_compiler is not None
+    assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    # verify that the output dataframe also has a populated relaxed query compiler
+    assert snow_result._query_compiler._relaxed_query_compiler is not None
+    assert (
+        snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    )
+
+    # create pandas dataframes
+    native_df = df.to_pandas()
+    native_df["C"] = native_df["B"] + 1
+    native_result = native_df
+
+    # compare results
+    assert_frame_equal(snow_result, native_result)
+
+
+@pytest.mark.parametrize(
+    "input_df2",
+    [
+        native_pd.DataFrame([[2, 112], [1, 111], [3, 113]], columns=["A", "B"]),
+        native_pd.DataFrame([[2, 112]], columns=["A", "B"]),
+    ],
+)
+@sql_count_checker(query_count=5, join_count=2)
+def test_set_2d_labels_from_different_df(session, input_df2):
+    # create tables
+    table_name1 = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.create_dataframe(
+        native_pd.DataFrame([[2, 12], [1, 11], [3, 13]], columns=["A", "B"])
+    ).write.save_as_table(table_name1, table_type="temp")
+
+    table_name2 = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.create_dataframe(input_df2).write.save_as_table(
+        table_name2, table_type="temp"
+    )
+
+    # create snow dataframes
+    df1 = pd.read_snowflake(table_name1)
+    df2 = pd.read_snowflake(table_name2)
+    # ensure both dataframes have same order
+    df1 = df1.sort_values("A", ignore_index=True)
+    df2 = df2.sort_values("A", ignore_index=True)
+    df3 = df1
+    df3["C"] = df2["B"] + 1
+    snow_result = df3
+
+    # verify that the input dataframe has a populated relaxed query compiler
+    assert df1._query_compiler._relaxed_query_compiler is not None
+    assert df1._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    # verify that the output dataframe also has a populated relaxed query compiler
+    assert snow_result._query_compiler._relaxed_query_compiler is not None
+    assert (
+        snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+    )
+
+    # create pandas dataframes
+    native_df1 = df1.to_pandas()
+    native_df2 = df2.to_pandas()
+    native_df3 = native_df1
+    native_df3["C"] = native_df2["B"] + 1
+    native_result = native_df3
 
     # compare results
     assert_frame_equal(snow_result, native_result)